requirements.txt
openai
python-dotenv
pyyaml
datasets